GH-462 Add placeholder api (#462)
* Add placeholder api

Create placeholder registry system
Create two sample placeholders

* "Refactor BridgeManager and PlaceholderApiExtension for dynamic naming

In BridgeManager, changed 'server' parameter to 'plugin' to reflect the requirement of PlaceholderApiExtension. In PlaceholderApiExtension, used 'plugin' to dynamically fetch the plugin name and version for getIdentifier() and getVersion(). This improved flexibility and reduced hardcoding. Unnecessary commented code in EternalCore was also removed."

* "Refactor BridgeManager and PlaceholderApiExtension for dynamic naming

In BridgeManager, changed 'server' parameter to 'plugin' to reflect the requirement of PlaceholderApiExtension. In PlaceholderApiExtension, used 'plugin' to dynamically fetch the plugin name and version for getIdentifier() and getVersion(). This improved flexibility and reduced hardcoding. Unnecessary commented code in EternalCore was also removed."

* Follow @Rollczi suggestions from code review.

* Correct naming convention in PlaceholderApiExtension

* Correct naming convention in BridgeManager

* Follow @Rollczi review.

---------

Co-authored-by: Martin Sulikowski <vLuckyyy.biznes@gmail.com>