GH-481 Add feature to support center-aligned text (#497)
* Refactor legacy classes location and add text centering feature

Relocated 'LegacyColorProcessor' and 'Legacy' classes from the 'util.legacy' package to 'adventure.legacy' to improve code organization as suggested in a code review. This required updating the import statements in multiple classes. Additionally, introduced a new feature to support center-aligned text. To implement this, added 'CenterTagResolver' class to 'core.adventure.resolver' package. To support this feature, 'pixel-width' libraries (core and utils) were added in the build file.

Co-Authored-By: Rafał Chomczyk <hello@rafal.moe>
Co-Authored-By: Martin Sulikowski <vLuckyyy.biznes@gmail.com>

* Rename constant in CenterTagResolver class

Refactored the constant variable name from "CENTER" to "TAG_NAME" in the CenterTagResolver.

* Simplify tag setting process in MiniMessage

The previous method of defining tags was unnecessarily complex and verbose, and it made code readability a challenge. The new method employs a lambda function for setting tags, inherently simplifying the structure and enhancing code readability. Also, a minor change was made in CenterTagResolver to access the 'component' instance directly for performance improvement as it avoids recurring unnecessary casting.

---------

Co-authored-by: Rafał Chomczyk <hello@rafal.moe>